top-machine-back-master/backend/middleware/auth.py
from datetime import datetime, timedelta, timezone

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials

from backend.config import settings

bearer_scheme = HTTPBearer()


def create_access_token(user_id: int) -> str:
    """Создать JWT токен для пользователя."""
    payload = {
        "sub": str(user_id),
        "exp": datetime.now(timezone.utc) + timedelta(days=settings.jwt_expire_days),
        "iat": datetime.now(timezone.utc),
    }
    return jwt.encode(payload, settings.jwt_secret, algorithm=settings.jwt_algorithm)


def decode_access_token(token: str) -> dict:
    """Декодировать и валидировать JWT токен."""
    try:
        payload = jwt.decode(
            token,
            settings.jwt_secret,
            algorithms=[settings.jwt_algorithm],
        )
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Токен истёк",
        )
    except jwt.InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Невалидный токен",
        )


def get_current_user_id(
    credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme),
) -> int:
    """
    Dependency для защищённых ручек.
    Извлекает user_id из токена.

    Использование:
        @router.post("/receive")
        async def receive_site(
            body: ReceiveSiteRequest,
            user_id: int = Depends(get_current_user_id),
        ):
            ...
    """
    payload = decode_access_token(credentials.credentials)
    return int(payload["sub"])
