top-machine-back-master/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
asyncpg==0.29.0
aiohttp==3.11.18
pydantic==2.7.1
pydantic-settings==2.3.1
PyJWT==2.8.0
dotenv==0.9.9
aiosmtplib==3.0.2
bcrypt==4.2.1
email-validator==2.2.0
httpx==0.27.0
aiohttp-socks==0.9.1
